Make HiveWriter a top level class
